+++ kernel/os/include/os/os_task.h
@@ -119,7 +119,7 @@ struct os_task {
     /** Total task run time */
     os_time_t t_run_time;
     /** Total task run time high resolution */
-    uint64_t t_run_time_hr;
+    uint32_t t_run_time_hr;
     /**
      * Total number of times this task has been context switched during
      * execution.
